Improving the resiliency of BigInt decoding.
It will now decode ints as well.

Note that this is just us being more tolerant of recieving something
like `1234`. If the number sent is actually in the "big" integer range
(>2^32) then it won't work and it's the sender's fault. The JSON spec
doesn't support numbers that big. They'd have to be sent as strings.